Refactor iqk: Factor out GEMM for iqk-quants (AVX2/AVX512)
